+++ mne/surface.py
@@ -16,10 +16,10 @@ from .fiff.constants import FIFF
 from .fiff.open import fiff_open
 from .fiff.tree import dir_tree_find
 from .fiff.tag import find_tag
-from .fiff.write import write_int, write_float, write_float_matrix, \
-                        write_int_matrix, start_file, end_block, \
-                        start_block, end_file, write_string, \
-                        write_float_sparse_rcs
+from .fiff.write import (write_int, write_float, write_float_matrix,
+                         write_int_matrix, start_file, end_block,
+                         start_block, end_file, write_string,
+                         write_float_sparse_rcs)
 from .utils import logger, verbose, get_subjects_dir
 
 #
@@ -203,6 +203,43 @@ def _read_bem_surface(fid, this, def_coord_frame, s_id=None):
     return res
 
 
+def fast_cross_3d(x, y):
+    """Compute cross product between list of 3D vectors
+
+    Much faster than np.cross() when the number of cross products
+    becomes large (>500). This is because np.cross() methods become
+    less memory efficient at this stage.
+
+    Parameters
+    ----------
+    x : array
+        Input array 1.
+    y : array
+        Input array 2.
+
+    Returns
+    -------
+    z : array
+        Cross product of x and y.
+
+    Notes
+    -----
+    x and y must both be 2D row vectors. One must have length 1, or both
+    lengths must match.
+    """
+    assert x.ndim == 2
+    assert y.ndim == 2
+    assert x.shape[1] == 3
+    assert y.shape[1] == 3
+    assert (x.shape[0] == 1 or y.shape[0] == 1) or x.shape[0] == y.shape[0]
+    if max([x.shape[0], y.shape[0]]) >= 500:
+        return np.c_[x[:, 1] * y[:, 2] - x[:, 2] * y[:, 1],
+                     x[:, 2] * y[:, 0] - x[:, 0] * y[:, 2],
+                     x[:, 0] * y[:, 1] - x[:, 1] * y[:, 0]]
+    else:
+        return np.cross(x, y)
+
+
 def _accumulate_normals(tris, tri_nn, npts):
     """Efficiently accumulate triangle normals"""
     # this code replaces the following, but is faster (vectorized):
@@ -255,7 +292,7 @@ def _complete_surface_info(this, do_neighbor_vert=False):
     r2 = this['rr'][this['tris'][:, 1], :]
     r3 = this['rr'][this['tris'][:, 2], :]
     this['tri_cent'] = (r1 + r2 + r3) / 3.0
-    this['tri_nn'] = np.cross((r2 - r1), (r3 - r1))
+    this['tri_nn'] = fast_cross_3d((r2 - r1), (r3 - r1))
 
     #   Triangle normals and areas
     size = np.sqrt(np.sum(this['tri_nn'] ** 2, axis=1))
@@ -433,6 +470,59 @@ def _normalize_vectors(rr):
     size = np.sqrt(np.sum(rr * rr, axis=1))
     size[size == 0] = 1.0  # avoid divide-by-zero
     rr /= size[:, np.newaxis]  # operate in-place
+
+
+def _compute_nearest(xhs, rr, use_balltree=True, return_dists=False):
+    """Find nearest neighbors
+
+    Note: The rows in xhs and rr must all be unit-length vectors, otherwise
+    the result will be incorrect.
+
+    Parameters
+    ----------
+    xhs : array, shape=(n_samples, n_dim)
+        Points of data set.
+    rr : array, shape=(n_query, n_dim)
+        Points to find nearest neighbors for.
+    use_balltree : bool
+        Use fast BallTree based search from scikit-learn. If scikit-learn
+        is not installed it will fall back to the slow brute force search.
+
+    Returns
+    -------
+    nearest : array, shape=(n_query,)
+        Index of nearest neighbor in xhs for every point in rr.
+    """
+    if use_balltree:
+        try:
+            from sklearn.neighbors import BallTree
+        except ImportError:
+            logger.info('Nearest-neighbor searches will be significantly '
+                        'faster if scikit-learn is installed.')
+            use_balltree = False
+
+    if use_balltree is True:
+        ball_tree = BallTree(xhs)
+        if return_dists:
+            out = ball_tree.query(rr, k=1, return_distance=True)
+            return out[1][:, 0], out[0][:, 0]
+        else:
+            nearest = ball_tree.query(rr, k=1, return_distance=False)[:, 0]
+            return nearest
+    else:
+        if return_dists:
+            nearest = list()
+            dists = list()
+            for r in rr:
+                d = cdist(r[np.newaxis, :], xhs)
+                idx = np.argmin(d)
+                nearest.append(idx)
+                dists.append(d[0, idx])
+            return (np.array(nearest), np.array(dists))
+        else:
+            nearest = np.array([np.argmin(cdist(r[np.newaxis, :], xhs))
+                                for r in rr])
+            return nearest
 
 
 def _get_nearest(to, fro):
@@ -899,7 +989,7 @@ def _make_morph_map(subject_from, subject_to, subjects_dir=None):
         c = np.sum(r12 * r13, axis=1)
         mat = np.rollaxis(np.array([[b, -c], [-c, a]]), 2)
         mat /= (a * b - c * c)[:, np.newaxis, np.newaxis]
-        tri_nn = np.cross(r12, r13)
+        tri_nn = fast_cross_3d(r12, r13)
 
         fname = op.join(subjects_dir, subject_to, 'surf',
                         '%s.sphere.reg' % hemi)