+++ doc/source/whats_new.rst
@@ -53,6 +53,8 @@ Changelog
 
    - Removed deprecated read/write_stc/w, use SourceEstimate methods instead
 
+   - Add volume and discrete source space creation support by `Eric Larson`_
+
 API
 ~~~